Move stuff around
- Collect errors in single file
- Put important functions on top, then logging functions, then builtins like __repr__
